move Index class to base.py, set __all__
