+++ src/fast_forward/ranking.py
@@ -81,7 +81,7 @@ class Ranking:
         :param name: Method name.
         :param queries: Query IDs mapped to queries.
         :param dtype: How the scores should be represented in the data frame.
-        :param copy:  Whether to copy the data frame.
+        :param copy: Whether to copy the data frame.
         :param is_sorted: Whether the data frame is already sorted (by score).
         :raises ValueError: When the queries are incomplete.
         """
