+++ src/fast_forward/util/__init__.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:
 
     import pandas as pd
 
-    from fast_forward.index import Index
+    from fast_forward.index.base import Index
     from fast_forward.ranking import Ranking
